Implementing property change notification
This version bubbles change notifications up from
the SimConfig to the configurator. Next round may
drop that and reintroduce equality testing so sim
can just use that to look for changes. Then, property
change notification may only have to be done on
transforms that will be changed programmatically
outside the GUI.